chore: split auth_1 test
